# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Prestige Motors Group API
  description: >
    Account ledger API for Prestige Motors Group. Covers authentication,
    transaction management, ledger account categorization, and reporting.


    Status legend used in each endpoint's description:
      - ✅ Implemented — live in the app right now
      - 🚧 Planned — documented here as the team's agreed contract, not built yet
  version: "0.1.0"
servers:
  - url: http://localhost:8080
    description: Local dev
...
security:
  - bearerAuth: []
paths:
...
  /api/transactions:
    get:
      tags: [Transactions]
      summary: List all transactions
      description: ✅ Implemented. Admin + Staff.
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
    post:
      tags: [Transactions]
      summary: Create a transaction
      description: ✅ Implemented. Admin + Staff.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
...
  /api/transactions/import:
    post:
      tags: [Transactions]
      summary: Bulk import transactions from a CSV file
      description: >
        🚧 Planned — CsvImportService is written, but not yet wired to this
        route. Admin only. File format: pipe-delimited
        `date|time|description|vendor|amount` per line.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Import summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
...
  /api/ledger-accounts:
    get:
      tags: [Ledger Accounts]
      summary: List all ledger accounts
      description: 🚧 Planned. Admin + Staff (read-only for staff).
      responses:
        '200':
          description: List of ledger accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LedgerAccount'
    post:
      tags: [Ledger Accounts]
      summary: Create a ledger account
      description: 🚧 Planned. Admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LedgerAccountInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerAccount'
  /api/ledger-accounts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    put:
      tags: [Ledger Accounts]
      summary: Update a ledger account
      description: 🚧 Planned. Admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LedgerAccountInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerAccount'
    delete:
      tags: [Ledger Accounts]
      summary: Delete a ledger account
      description: >
        🚧 Planned. Admin only. Team decision needed: reassign existing
        transactions to "Uncategorized" first, or block delete if any
        transactions still reference this account.
      responses:
        '204':
          description: Deleted
        '409':
          description: Conflict — transactions still reference this account
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Paste the `token` value returned from POST /login here (Swagger's
        "Authorize" button, or Postman's Bearer Token auth type).
  schemas:
...
    LedgerAccountInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Repair & Mechanic
        description:
          type: string
          example: Vehicle repair and maintenance costs
    LedgerAccount:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              readOnly: true
        - $ref: '#/components/schemas/LedgerAccountInput'
    TransactionInput:
      type: object
      required: [date, time, description, vendor, amount]
      properties:
        date:
          type: string
          format: date
          example: "2026-07-13"
        time:
          type: string
          example: "14:30:00"
        description:
          type: string
          example: Oil change service
        vendor:
          type: string
          example: QuickLube Auto
        amount:
          type: number
          format: double
          example: -89.99
        ledgerAccountId:
          type: integer
          nullable: true
          description: FK to a LedgerAccount; null/omitted = Uncategorized
    Transaction:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              readOnly: true
        - $ref: '#/components/schemas/TransactionInput'
        - type: object
          properties:
            ledgerAccount:
              $ref: '#/components/schemas/LedgerAccount'
              nullable: true
    ImportResult:
      type: object
      properties:
        importedCount:
          type: integer
        errors:
          type: array
          items:
            type: string
...